+++ app/controllers/api_v1/measures_controller.rb
@@ -167,13 +167,13 @@ module ApiV1
         return
       end
 
-      begin
-        @calculator_options = { doPretty: true }
-        calculated_results = BonnieBackendCalculator.calculate(@api_v1_measure, @api_v1_patients, @api_v1_value_sets, @calculator_options)
-      rescue StandardError
-        render json: {status: "error", messages: "Error calculating the results."}, status: :internal_server_error
-        return
-      end
+      # begin
+      @calculator_options = { doPretty: true }
+      calculated_results = BonnieBackendCalculator.calculate(@api_v1_measure, @api_v1_patients, @api_v1_value_sets, @calculator_options)
+      # rescue StandardError
+        # render json: {status: "error", messages: "Error calculating the results."}, status: :internal_server_error
+        # return
+      # end
 
       if request.headers['Accept'] == Mime::Type.lookup_by_extension(:xlsx)
         begin
